+++ plugin-opencode/examples/.opencode/plugins/openlogos-local.js
@@ -1,0 +1,4 @@
+import { OpenLogosPlugin } from "../../../../src/index.js";
+
+export { OpenLogosPlugin };
+export default OpenLogosPlugin;